+++ src/historySession.js
@@ -1,0 +1,3 @@
+import reducer from './reducer'
+// Send to standard reducer first.
+// If a.